+++ CHANGELOG.md
@@ -1,7 +1,15 @@
+Version: **1.0.1**
+
+Date Released: **12/12/2025**
+
+- Static analysis support for user accounts
+- Improved `.json` output for detector results
+
 Version: **1.0.0**
 
 Date Released: **11/30/2025**
 
 - Initial FirmaForge release
 - Functional detector and extractor with limited firmware support
+- Restricted to OpenWrt firmware
 - Docker container support